#113[php,controllers] 1-create assigen manger quiers 3-validatethe assigend data 4-show error

+++ database/migrations/2019_02_04_210348_create_inventories_table.php
@@ -15,7 +15,7 @@ class CreateInventoriesTable extends Migration
     {
         Schema::create('inventories', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedTinyInteger('serial_code');
+            $table->integer('serial_code');
             $table->string('storage');
             $table->string('country');
             $table->string('city');
